+++ sndr/product_api/legacy/web_static/assets/arrow-up-DLl1klp8.js
@@ -1,0 +1,1 @@
+import{f as e}from"./api-B3vUbj-J.js";var t=e(`ArrowUp`,[[`path`,{d:`m5 12 7-7 7 7`,key:`hav0vg`}],[`path`,{d:`M12 19V5`,key:`x0mq9r`}]]);export{t};